feat(admin-console): move topic subscriptions to relay page
