+++ docs/01_project/activeContext/tasks/completed/2026-02-24.md
@@ -18,3 +18,25 @@
 - [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job format-check`（pass）
 - [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job native-test-linux`（pass）
 - [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job community-node-tests`（pass）
+
+## Issue #153: Community Node Admin UI に Relay サイドバーを追加し、トピック購読管理を移設
+
+- [x] `kukuri-community-node/apps/admin-console/src/App.tsx` のサイドバーに `Relay` 項目を追加した。
+- [x] `kukuri-community-node/apps/admin-console/src/router.tsx` に `/relay` ルートを追加した。
+- [x] `kukuri-community-node/apps/admin-console/src/pages/SubscriptionsPage.tsx` からトピック購読管理 UI を除去し、購読申請・プラン・利用状況管理に責務を整理した。
+- [x] `kukuri-community-node/apps/admin-console/src/pages/RelayPage.tsx` を新規作成し、トピック購読 CRUD と ingest policy 管理を実装した。
+- [x] `kukuri-community-node/crates/cn-admin-api/src/subscriptions.rs` に topic ごとの `connected_users` / `connected_user_count` 集計を追加し、node subscription レスポンスへ反映した。
+- [x] `kukuri-community-node/crates/cn-admin-api/src/contract_tests.rs` を更新し、接続ユーザー数・一覧の契約テストを追加した。
+- [x] `kukuri-community-node/apps/admin-console/src/pages/RelayPage.test.tsx` を新規作成し、CRUD と接続ユーザー表示のテストを追加した。
+- [x] 進捗レポート `docs/01_project/progressReports/2026-02-24_issue153_relay_sidebar_topic_subscriptions.md` を追加した。
+
+## 検証（Issue #153）
+
+- [x] `docker compose -f docker-compose.test.yml up -d community-node-postgres`（pass）
+- [x] `docker compose -f docker-compose.test.yml build test-runner`（pass）
+- [x] `docker run --rm --network kukuri_community-node-network -e DATABASE_URL=postgres://cn:cn_password@community-node-postgres:5432/cn -v "$(git rev-parse --show-toplevel):/workspace" -w /workspace/kukuri-community-node kukuri-test-runner bash -lc "set -euo pipefail; source /usr/local/cargo/env; cargo run --locked -p cn-cli -- openapi export --service user-api --output apps/admin-console/openapi/user-api.json --pretty; cargo run --locked -p cn-cli -- openapi export --service admin-api --output apps/admin-console/openapi/admin-api.json --pretty"`（pass）
+- [x] `docker run --rm --network kukuri_community-node-network -e DATABASE_URL=postgres://cn:cn_password@community-node-postgres:5432/cn -v "$(git rev-parse --show-toplevel):/workspace" -w /workspace/kukuri-community-node/apps/admin-console kukuri-test-runner bash -lc "set -euo pipefail; source /usr/local/cargo/env; pnpm install --frozen-lockfile; pnpm generate:api; pnpm typecheck; pnpm test -- src/App.test.tsx src/pages/RelayPage.test.tsx src/pages/SubscriptionsPage.test.tsx"`（pass）
+- [x] `docker run --rm --network kukuri_community-node-network -e DATABASE_URL=postgres://cn:cn_password@community-node-postgres:5432/cn -v "$(git rev-parse --show-toplevel):/workspace" -w /workspace/kukuri-community-node kukuri-test-runner bash -lc "set -euo pipefail; source /usr/local/cargo/env; cargo test -p cn-admin-api subscription_requests_and_node_subscriptions_contract_success -- --nocapture; cargo test --workspace --all-features; cargo build --release -p cn-cli; cargo fmt --all --check"`（pass）
+- [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job format-check`（pass）
+- [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job native-test-linux`（pass）
+- [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job community-node-tests`（pass）
